sapt lgtm changes

+++ psi4/src/psi4/libsapt_solver/ind22.cc
@@ -145,7 +145,7 @@ double SAPT2::ind202() {
 }
 
 double SAPT2::ind220_1(int intfile, const char *AAlabel, const char *ARlabel, const char *RRlabel, int ampfile,
-                       const char *tlabel, double **iAR, double **wBAA, double **wBRR, int foccA, int noccA, int nvirA,
+                       const char *tlabel, double **iAR, double **wBAA, double **wBRR, size_t foccA, size_t noccA, size_t nvirA,
                        double *evalsA) {
     int aoccA = noccA - foccA;
 
@@ -213,9 +213,9 @@ double SAPT2::ind220_1(int intfile, const char *AAlabel, const char *ARlabel, co
     return (energy);
 }
 
-double SAPT2::ind220_2(int ampfile, const char *tlabel, double **iAR, double **wBAA, double **wBRR, int foccA,
-                       int noccA, int nvirA) {
-    int aoccA = noccA - foccA;
+double SAPT2::ind220_2(int ampfile, const char *tlabel, double **iAR, double **wBAA, double **wBRR, size_t foccA,
+                       size_t noccA, size_t nvirA) {
+    size_t aoccA = noccA - foccA;
 
     double **tAR = block_matrix(aoccA, nvirA);
     psio_->read_entry(ampfile, tlabel, (char *)tAR[0], sizeof(double) * aoccA * nvirA);
@@ -238,9 +238,9 @@ double SAPT2::ind220_2(int ampfile, const char *tlabel, double **iAR, double **w
     return (energy);
 }
 
-double SAPT2::ind220_3(int ampfile, const char *AAlabel, const char *RRlabel, double **iAR, double **wBAR, int foccA,
-                       int noccA, int nvirA) {
-    int aoccA = noccA - foccA;
+double SAPT2::ind220_3(int ampfile, const char *AAlabel, const char *RRlabel, double **iAR, double **wBAR, size_t foccA,
+                       size_t noccA, size_t nvirA) {
+    size_t aoccA = noccA - foccA;
 
     double **pAA = block_matrix(aoccA, aoccA);
     double **pRR = block_matrix(nvirA, nvirA);
@@ -271,9 +271,9 @@ double SAPT2::ind220_3(int ampfile, const char *AAlabel, const char *RRlabel, do
     return (energy);
 }
 
-double SAPT2::ind220_4(int ampfile, const char *thetalabel, int intfile, const char *ARlabel, double **iAR, int foccA,
-                       int noccA, int nvirA) {
-    int aoccA = noccA - foccA;
+double SAPT2::ind220_4(int ampfile, const char *thetalabel, int intfile, const char *ARlabel, double **iAR, size_t foccA,
+                       size_t noccA, size_t nvirA) {
+    size_t aoccA = noccA - foccA;
 
     double **xAA = block_matrix(aoccA, aoccA);
     double **xRR = block_matrix(nvirA, nvirA);
@@ -311,8 +311,8 @@ double SAPT2::ind220_4(int ampfile, const char *thetalabel, int intfile, const c
     return (energy);
 }
 
-double SAPT2::ind220_5(int ampfile, const char *tlabel, double **iAR, int foccA, int noccA, int nvirA, double *evalsA) {
-    int aoccA = noccA - foccA;
+double SAPT2::ind220_5(int ampfile, const char *tlabel, double **iAR, size_t foccA, size_t noccA, size_t nvirA, double *evalsA) {
+    size_t aoccA = noccA - foccA;
 
     double **tARAR = block_matrix(aoccA * nvirA, aoccA * nvirA);
     psio_->read_entry(ampfile, tlabel, (char *)tARAR[0], sizeof(double) * aoccA * nvirA * aoccA * nvirA);
@@ -345,7 +345,7 @@ double SAPT2::ind220_5(int ampfile, const char *tlabel, double **iAR, int foccA,
 }
 
 double SAPT2::ind220_6(int intfile, const char *AAlabel, const char *ARlabel, const char *RRlabel, int ampfile,
-                       const char *tlabel, double **iAR, int foccA, int noccA, int nvirA) {
+                       const char *tlabel, double **iAR, size_t foccA, size_t noccA, size_t nvirA) {
     int aoccA = noccA - foccA;
 
     double **B_p_AR = get_DF_ints(intfile, ARlabel, foccA, noccA, 0, nvirA);
@@ -398,9 +398,9 @@ double SAPT2::ind220_6(int intfile, const char *AAlabel, const char *ARlabel, co
 
 double SAPT2::ind220_7(int AAfile, const char *AAlabel, const char *ARlabel, const char *RRlabel, int BBfile,
                        const char *BSlabel, int ampfile, const char *tlabel, const char *pAAlabel, const char *pRRlabel,
-                       double **iBS, int foccA, int noccA, int nvirA, int foccB, int noccB, int nvirB) {
-    int aoccA = noccA - foccA;
-    int aoccB = noccB - foccB;
+                       double **iBS, size_t foccA, size_t noccA, size_t nvirA, size_t foccB, size_t noccB, size_t nvirB) {
+    size_t aoccA = noccA - foccA;
+    size_t aoccB = noccB - foccB;
 
     double **pAA = block_matrix(aoccA, aoccA);
     double **tAR = block_matrix(aoccA, nvirA);
