+++ psi4/src/psi4/libsapt_solver/sapt2.h
@@ -105,10 +105,10 @@ class SAPT2 : public SAPT {
 
     double **get_DF_ints(int, const char *, int, int, int, int);
     double **get_DF_ints_nongimp(int, const char *, int, int, int, int);
-    void antisym(double *, int, int);
-    void antisym(double **, int, int);
+    void antisym(double *, size_t, size_t);
+    void antisym(double **, size_t, size_t);
 
-    void cphf_solver(double **, double **, double *, int, const char *, const char *, const char *, int, int);
+    void cphf_solver(double **, double **, double *, int, const char *, const char *, const char *, size_t, size_t);
 
     void exch_ind20rA_B();
     void exch_ind20rB_A();
@@ -133,10 +133,10 @@ class SAPT2 : public SAPT {
     void ijkl_to_ikjl(double *, int, int, int, int);
     void symmetrize(double *, int, int);
 
-    void natural_orbitalify(int, const char *, double *evals, int, int, int, const char);
+    void natural_orbitalify(int, const char *, double *evals, int, int, size_t, const char);
     void natural_orbitalify_df_ints();
 
-    double elst120(double **, double **, double **, int, const char *, const char *, const char *, int, int, int);
+    double elst120(double **, double **, double **, int, const char *, const char *, const char *, size_t, size_t, size_t);
 
     double exch110(int, const char *);
     double exch101(int, const char *);
@@ -159,14 +159,14 @@ class SAPT2 : public SAPT {
     double ind220();
     double ind202();
     double ind220_1(int, const char *, const char *, const char *, int, const char *, double **, double **, double **,
-                    int, int, int, double *);
-    double ind220_2(int, const char *, double **, double **, double **, int, int, int);
-    double ind220_3(int, const char *, const char *, double **, double **, int, int, int);
-    double ind220_4(int, const char *, int, const char *, double **, int, int, int);
-    double ind220_5(int, const char *, double **, int, int, int, double *);
-    double ind220_6(int, const char *, const char *, const char *, int, const char *, double **, int, int, int);
+                    size_t, size_t, size_t, double *);
+    double ind220_2(int, const char *, double **, double **, double **, size_t, size_t, size_t);
+    double ind220_3(int, const char *, const char *, double **, double **, size_t, size_t, size_t);
+    double ind220_4(int, const char *, int, const char *, double **, size_t, size_t, size_t);
+    double ind220_5(int, const char *, double **, size_t, size_t, size_t, double *);
+    double ind220_6(int, const char *, const char *, const char *, int, const char *, double **, size_t, size_t, size_t);
     double ind220_7(int, const char *, const char *, const char *, int, const char *, int, const char *, const char *,
-                    const char *, double **, int, int, int, int, int, int);
+                    const char *, double **, size_t, size_t, size_t, size_t, size_t, size_t);
 
    public:
     SAPT2(SharedWavefunction Dimer, SharedWavefunction MonomerA, SharedWavefunction MonomerB, Options &options,
