+++ psi4/src/psi4/libsapt_solver/utils.cc
@@ -1323,7 +1323,7 @@ double **SAPT2::get_DF_ints_nongimp(int filenum, const char *label, int startA, 
     return AA;
 }
 
-void SAPT2::antisym(double *A, int nocc, int nvir) {
+void SAPT2::antisym(double *A, size_t nocc, size_t nvir) {
     double *X = init_array(nvir);
 
     for (int a = 0; a < nocc; a++) {
@@ -1345,7 +1345,7 @@ void SAPT2::antisym(double *A, int nocc, int nvir) {
     free(X);
 }
 
-void SAPT2::antisym(double **A, int nocc, int nvir) {
+void SAPT2::antisym(double **A, size_t nocc, size_t nvir) {
     double *X = init_array(nvir);
 
     for (int a = 0; a < nocc; a++) {
